+++ Diamond Engine/Source Code/WI_Console.cpp
@@ -24,7 +24,7 @@ void W_Console::Draw()
 		ImGui::SetCursorPosX(offset);
 		if (ImGui::Button("Clear"/*, ImGuiDir_Right*/))
 		{
-			for (int i = 0; i < logs.size(); i++)
+			for (size_t i = 0; i < logs.size(); i++)
 			{
 				logs[i].msg.clear();
 			}
